New Commands: SAVEINT, MAP, EXIT, GETFOLDERNAMES, NEXTFOLDER. New commands in omacam: VIDOPENFILE, VIDCLOSEFILE, VIDADDFRAME, CVHOUGHCIRCLES. %x and %y now work properly. %k gives date as YYYY-MM-DD. dcrawx has fewer warnings/errors using std=c++11 omacam needs opencv installed if OPENCV_ROUTINES is defined in cameraSelector.h

+++ oma2/oma2/sbig.mm
@@ -64,7 +64,7 @@ void setReturnValues(){
     values[EXPOSURE] = ExposureTime;
     iBuffer.setvalues(values);
     free(values);
-    float extraValues[4]={setpointTemp,ccdTemp,ExposureTime,Filter};
+    float extraValues[4]={setpointTemp,ccdTemp,ExposureTime,(float)Filter};
     iBuffer.setExtra(extraValues,4);
 }
